# Optimize cpu sketch allreduce for sparse data.