Follow up on https://github.com/dmlc/xgboost/pull/5880 . @ShvetsKS @SmirnovEgorRu   Previously sketching on URL dataset on distributed environment was not possible due to memory usage blow up.  After this PR I can unify the sketching for `hist` and `approx` and proceed with other refactoring to unify the codebase.

Also fixed a bug in distributed training with incorrect allreduce call, and added tests.

# Perf
Previous perf is from #5880.

## Single Node
### Before
- URL
GmatInitialization: 10.3033s, 1 calls @ 10303323us
GmatInitialization: 10.2406s, 1 calls @ 10240576us
GmatInitialization: 10.224s, 1 calls @ 10224020us

- HIGGS
GmatInitialization: 3.74322s, 1 calls @ 3743224us
GmatInitialization: 3.63079s, 1 calls @ 3630793us
GmatInitialization: 3.67905s, 1 calls @ 3679049us


### After
- HIGGS
GmatInitialization: 3.7794s, 1 calls @ 3779401us
GmatInitialization: 3.81085s, 1 calls @ 3810851us
GmatInitialization: 3.77258s, 1 calls @ 3772581us
GmatInitialization: 3.76844s, 1 calls @ 3768436us

- URL
GmatInitialization: 10.315s, 1 calls @ 10314976us
GmatInitialization: 10.3202s, 1 calls @ 10320246us
GmatInitialization: 10.3059s, 1 calls @ 10305877us

## Multi Node 4x4
### Before
- HIGGS
GmatInitialization: 3.7682s, 1 calls @ 3768198us
GmatInitialization: 3.707s, 1 calls @ 3707000us
GmatInitialization: 3.70828s, 1 calls @ 3708276us

- URL
NAN


### After
- HIGGS
GmatInitialization: 3.64623s, 1 calls @ 3646232us
GmatInitialization: 3.5942s, 1 calls @ 3594197us
GmatInitialization: 3.73079s, 1 calls @ 3730792us

- URL
GmatInitialization: 6.53294s, 1 calls @ 6532936us
GmatInitialization: 6.43215s, 1 calls @ 6432153us
GmatInitialization: 6.67619s, 1 calls @ 6676192us


@ShvetsKS Also I might have found the cause of the slow down you mentioned, this is memory usage from massif when training on URL on distributed env, but should be related to performance:
![Screenshot from 2020-08-13 05-15-24](https://user-images.githubusercontent.com/16746409/90069609-1db2b180-dd25-11ea-96a6-9dd9453ee4ed.png)

I can't get a complete run as my home machine only has 32 GB of memory.  Hope that helps.