@trivialfis 2020-08-17
Also fixed a bug in distributed training with incorrect allreduce call.