@trivialfis 2020-08-18
@ShvetsKS I'm merging it as it contains a bug fix that's blocking the CI.  If you have concerns feel free to revert or let me know how should I make follow up PRs.